Summary New Picklist Trees whee
Implemented MenuItemPicklist to replace MenuItemSwap
Keeping MenuItemSwap for reference.

Behavior:
If a menu item on the Other Menus picklist is selected and the Right Button is clicked, the Menu Item will be added to the Master Menu's MenuItem list.

If a menu item on the Master Menu picklist is selected, and a Menu on the Other Menus picklist is selected, and the Left Button is clicked, the Menu Item will be added to the array for the Selected Menu.